+++ src/features/course-main/course-main.module.scss
@@ -1,26 +1,28 @@
 .container {
-  @include media-laptop {
-    padding: 100px 40px;
-  }
-
-  display: flex;
-  gap: 40px;
-  align-items: center;
-  justify-content: flex-start;
-
-  padding: 100px;
-
-  text-align: left;
-
   background: $color-secondary-yellow;
 
-  .icon {
-    @include media-tablet {
-      display: none;
+  .content {
+    @include media-laptop {
+      padding: 100px 40px;
     }
+    
+    display: flex;
+    gap: 40px;
+    align-items: center;
+    justify-content: flex-start;
 
-    min-width: 150px;
-    max-width: 200px;
-    margin-top: -36px;
+    padding: 100px;
+
+    text-align: left;
+
+    .icon {
+      @include media-tablet {
+        display: none;
+      }
+
+      min-width: 150px;
+      max-width: 200px;
+      margin-top: -36px;
+    }
   }
 }
